+++ src/main/java/edu/ie3/simona/api/data/connection/ExtResultDataConnection.java
@@ -11,88 +11,49 @@ import edu.ie3.simona.api.ontology.results.ProvideResultEntities;
 import edu.ie3.simona.api.ontology.results.RequestResultEntities;
 import edu.ie3.simona.api.ontology.results.ResultDataMessageFromExt;
 import edu.ie3.simona.api.ontology.results.ResultDataResponseMessageToExt;
-
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
-import java.util.stream.Stream;
+import java.util.stream.Collectors;
 
 /** Enables data connection of results between SIMONA and SimonaAPI. */
 public final class ExtResultDataConnection
     extends BiDirectional<ResultDataMessageFromExt, ResultDataResponseMessageToExt> {
 
-  /** Map uuid to external id of grid related entities. */
-  private final List<UUID> gridResults;
+  /** A list of uuids for all registered result entities. */
+  private final List<UUID> resultUuids;
 
-  /** Map uuid to external id of system participants. */
-  private final List<UUID> participantResults;
-
-  /** Map uuid to external id of participant flex options. */
-  private final List<UUID> flexResults;
-
-  public ExtResultDataConnection(
-      List<UUID> participantResults, List<UUID> gridResults, List<UUID> flexResults) {
-    this.participantResults = participantResults;
-    this.gridResults = gridResults;
-    this.flexResults = flexResults;
+  public ExtResultDataConnection(List<UUID> results) {
+    this.resultUuids = results;
   }
 
-  public List<UUID> getGridResultDataAssets() {
-    return gridResults;
-  }
-
-  public List<UUID> getParticipantResultDataAssets() {
-    return participantResults;
-  }
-
-  public List<UUID> getFlexOptionAssets() {
-    return flexResults;
+  public List<UUID> getResultUuids() {
+    return Collections.unmodifiableList(resultUuids);
   }
 
   /** Method for requesting SIMONA results as list from an external simulation. */
   private List<ResultEntity> requestResultList(long tick) throws InterruptedException {
-    return createResultList(requestResults(tick));
+    sendExtMsg(new RequestResultEntities(tick, resultUuids));
+    return receiveWithType(ProvideResultEntities.class).results();
   }
 
-  private List<ResultEntity> requestFlexOptionResultsList(long tick) throws InterruptedException {
-    return createResultList(requestFlexOptionResults(tick));
-  }
-
-  private List<ResultEntity> requestGridResultsList(long tick) throws InterruptedException {
-    return createResultList(requestGridResults(tick));
-  }
-
-  private List<ResultEntity> requestParticipantResultsList(long tick) throws InterruptedException {
-    return createResultList(requestParticipantResults(tick));
+  private List<ResultEntity> requestResultList(long tick, List<UUID> entities)
+      throws InterruptedException {
+    sendExtMsg(new RequestResultEntities(tick, entities));
+    return receiveWithType(ProvideResultEntities.class).results();
   }
 
   /** Method for requesting SIMONA results as a map uuid to object from an external simulation. */
-  public Map<UUID, List<ResultEntity>> requestResults(long tick) throws InterruptedException {
-    List<UUID> allExtEntities =
-            Stream.concat(
-                            Stream.concat(getFlexOptionAssets().stream(), getGridResultDataAssets().stream()),
-                            getParticipantResultDataAssets().stream())
-                    .toList();
-    sendExtMsg(new RequestResultEntities(tick, allExtEntities));
-    return receiveWithType(ProvideResultEntities.class).results();
+  public Map<UUID, ResultEntity> requestResults(long tick) throws InterruptedException {
+    return createResultMap(requestResultList(tick));
   }
 
-  public Map<UUID, List<ResultEntity>> requestFlexOptionResults(long tick) throws InterruptedException {
-    sendExtMsg(new RequestResultEntities(tick, getFlexOptionAssets()));
-    return receiveWithType(ProvideResultEntities.class).results();
-  }
+    public Map<UUID, ResultEntity> requestResults(long tick, List<UUID> entities) throws InterruptedException {
+        return createResultMap(requestResultList(tick, entities));
+    }
 
-  public Map<UUID, List<ResultEntity>> requestGridResults(long tick) throws InterruptedException {
-    sendExtMsg(new RequestResultEntities(tick, getGridResultDataAssets()));
-    return receiveWithType(ProvideResultEntities.class).results();
-  }
-
-  public Map<UUID, List<ResultEntity>> requestParticipantResults(long tick) throws InterruptedException {
-    sendExtMsg(new RequestResultEntities(tick, getParticipantResultDataAssets()));
-    return receiveWithType(ProvideResultEntities.class).results();
-  }
-
-  private List<ResultEntity> createResultList(Map<UUID, List<ResultEntity>> results) {
-    return results.values().stream().flatMap(List::stream).toList();
+  private Map<UUID, ResultEntity> createResultMap(List<ResultEntity> results) {
+    return results.stream().collect(Collectors.toMap(ResultEntity::getInputModel, i -> i));
   }
 }