+++ src/main/java/edu/ie3/simona/api/mapping/DataType.java
@@ -6,31 +6,20 @@
 
 package edu.ie3.simona.api.mapping;
 
-import edu.ie3.datamodel.exceptions.ParsingException;
-
+/** Data types. */
 public enum DataType {
-  EXT_PRIMARY_INPUT("primary_input"),
-  EXT_EM_INPUT("em_input"),
-  EXT_EM_COMMUNICATION("em_communication"),
-  EXT_GRID_RESULT("grid_result"),
-  EXT_PARTICIPANT_RESULT("participant_result"),
-  EXT_FLEX_OPTIONS_RESULT("flex_options_result");
+  PRIMARY,
+  RESULT,
+  PRIMARY_RESULT,
+  EM;
 
-  public final String type;
-
-  DataType(String type) {
-    this.type = type;
+  /** Returns all primary types. */
+  public static DataType[] primaryTypes() {
+    return new DataType[] {PRIMARY, PRIMARY_RESULT};
   }
 
-  public static DataType parse(String type) throws ParsingException {
-    return switch (type) {
-      case "primary_input" -> EXT_PRIMARY_INPUT;
-      case "em_input" -> EXT_EM_INPUT;
-      case "em_communication" -> EXT_EM_COMMUNICATION;
-      case "grid_result" -> EXT_GRID_RESULT;
-      case "participant_result" -> EXT_PARTICIPANT_RESULT;
-      case "flex_options_result" -> EXT_FLEX_OPTIONS_RESULT;
-      default -> throw new ParsingException("Data type " + type + " is not supported!");
-    };
+  /** Returns all result types. */
+  public static DataType[] resultTypes() {
+    return new DataType[] {RESULT, PRIMARY_RESULT};
   }
 }